src/requirements/base.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
faiss-gpu==1.7.2
filelock==3.16.1
firecrawl-py==1.2.4
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.23
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langsmith==0.1.147
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.57.1
orjson==3.10.12
packaging==24.1
parsel==1.9.1
pillow==11.0.0
propcache==0.2.1
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
Scrapy==2.11.2
sentence-transformers==3.3.1
service-identity==24.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tldextract==5.1.2
tokenizers==0.21.0
tomli==2.0.2
torch==2.5.1
tqdm==4.66.5
transformers==4.47.0
triton==3.1.0
Twisted==24.7.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
websockets==13.1
yarl==1.18.3
zope.interface==7.0.3
